Add ability to get world from Entity directly
This allows the entity to alter itself through the world. It allows the
entity to remove itself from the world or move to another world. Calling
the kill() function on LivingEntity's now triggers remove() on the Entity,
which forces the world to remove and delete the entity. This is just for
added flexibility, especially if the Entity ID is lost and the user is
uncertain as to what world the entity is in.